Add NoBangStateMachineEvents Cop
State machine parses symbols that are passed as event names and defines
some methods. For example, given event named `ship` it will define
`ship` that will fail gracefully by returning false and `ship!` that
will raise an error.

Unfortunately, should you decide to pass a symbol that already ends with
a `!` state_machine will define `ship!` and `ship!!`.
That causes not only confusion, but also bugs and we have a
bunch of occurrences of this in our codebase hence this cop