RF-17561 Add rubocop-rails so we can use the Rails-specific cops
I've added it in such a way that it can be enabled for Rails projects
without it being forced upon unsuspecting Ruby-only projects.

I haven't actually gone through all the cops yet to decide what should
be on/off - I figure it'll be easier to see what the recommendations are
and go from there.

And I decided to bump the version up to 0.5.0 since this is adding a
whole new set of cops - I thought that warranted more than a patch
increase.